+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/Utils/ObjectAsMessage.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symfony\Component\DependencyInjection\Tests\Fixtures\Utils;
+
+use Symfony\Component\Messenger\Attribute\AsMessage;
+
+#[AsMessage]
+class ObjectAsMessage
+{
+}